requirements.txt
spacy==3.7.2
pdfplumber==0.10.3
pyyaml==6.0.1
watchdog==3.0.0
requests==2.31.0